Move Badge related services to dedicated file
